Get it working with a standalone "pot" server
Most of the barrier to this was my lack of understanding about how the pieces
fit together, but we also needed to get more services on the shared
passwordless SSH network, and also set up more services with a correct
"still.cfg" file so that RTP commands could be invoked.

I also changed the default Librarian data directory to /data for consistency
with everything else. One has to be careful with this stuff though since the
Librarian code hardcodes some questionable filepath manipulations.