Prepare for testing "onsite"/"offsite" librarian setup.
To simulate the full system, we'll want to run multiple Librarians and pot
servers, simulating a set of computers that are on the array site and a batch
that are off it. Get started on that by creating multiple Librarian databases
and prefixing various names that are going to be duplicated.

We also just set up all "hera-stack"-based images with preloaded hl_client.cfg
files that describe things how we want. No reason not to.