test-rtp: wait for database availability before starting up
If the database isn't available when we start, we just crash instantly, which
is bad. And Docker doesn't really know when the database is up since it
(currently) isn't able to wait and see when the network port is finally
actually open for connections. So we just manually wait until the port
is available before starting up the script.

The Librarian has similar needs, but it doesn't actually *do* anything until
someone connects to the Apache server, and each DB connection is short-lived.
So I think it makes more sense to leave the Librarian as-is.